Renamed default and automatic assemblers
