+++ core/src/test/java/org/seedstack/business/core/interfaces/assembler/dsl/fixture/customer/AutoAssembler.java
@@ -10,12 +10,12 @@
 package org.seedstack.business.core.interfaces.assembler.dsl.fixture.customer;
 
 import org.modelmapper.ModelMapper;
-import org.seedstack.business.core.interfaces.AutomaticAssembler;
+import org.seedstack.business.core.interfaces.ModelMapperAssembler;
 
 /**
 * @author pierre.thirouin@ext.mpsa.com (Pierre Thirouin)
 */
-public class AutoAssembler extends AutomaticAssembler<Order, OrderDto> {
+public class AutoAssembler extends ModelMapperAssembler<Order, OrderDto> {
     @Override
     protected ModelMapper configureAssembly() {
         return new ModelMapper();
